Trait libra_sumcheck::data_structures::random::FeedableRNG

pub trait FeedableRNG: RngCore {
    type Error: Error + From<Error>;
    fn setup() -> Self;

    fn feed<M: ToBytes>(&mut self, msg: &M) -> Result<(), Self::Error>;
}

Random Field Element Generator where randomness feed adds entropy for the output.

Implementation should support all types of input that has ToBytes trait.

Same sequence of feed and get call should yield same result!

Associated Types

type Error: Error + From<Error>

Error type

Loading content...
Required methods

fn setup() -> Self

Setup should not have any parameter.

fn feed<M: ToBytes>(&mut self, msg: &M) -> Result<(), Self::Error>

The feed message provide randomness for the generator, given the message.

Loading content...