Module libra_sumcheck::gkr_func_sumcheck

prover & verifier protocol for gkr function Interactive Protocol for GKR Function

Modules

libra
Implementation of GKR function by Xie et al.

Traits

GKRFuncProof
proof of the GKRFuncTheorem

GKRFuncSubClaim
subclaim: point and evaluation of that GKR

GKRFuncSumcheck
Sumcheck protocol for GKR Functions

GKRFuncTheorem
fixed point g, and the sum of the GKR function.